# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "装修预算与支出管理 API 服务，提供预算编制、支出审批、供应商管理与对账结算能力。",
        "title": "Renovation Budget API",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:19306",
    "basePath": "/api/v1",
    "paths": {
...
        "/budget-adjustments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "budget-adjustments"
                ],
                "summary": "查询预算调整单列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "预算表ID",
                        "name": "budget_sheet_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "审批状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    }
                }
            }
        },
        "/budget-adjustments/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "budget-adjustments"
                ],
                "summary": "获取预算调整单详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "调整单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    }
                }
            }
        },
        "/budget-adjustments/{id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "budget-adjustments"
                ],
                "summary": "审批通过预算调整单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "调整单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "审批请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ApproveBudgetAdjustmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    }
                }
            }
        },
        "/budget-adjustments/{id}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "budget-adjustments"
                ],
                "summary": "驳回预算调整单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "调整单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "驳回请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RejectBudgetAdjustmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    }
                }
            }
        },
...
        "/budgets/{id}/adjustments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "budget-adjustments"
                ],
                "summary": "查询指定预算表的调整单列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "预算表ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "审批状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "budget-adjustments"
                ],
                "summary": "提交预算调整单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "预算表ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "调整单请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateBudgetAdjustmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ApproveBudgetAdjustmentRequest": {
            "type": "object",
            "properties": {
                "approval_comment": {
                    "type": "string",
                    "maxLength": 512
                }
            }
        },
...
        "dto.CreateBudgetAdjustmentRequest": {
            "type": "object",
            "required": [
                "reason",
                "total_amount"
            ],
            "properties": {
                "reason": {
                    "type": "string",
                    "maxLength": 512
                },
                "total_amount": {
                    "type": "number"
                }
            }
        },
...
        "dto.RejectBudgetAdjustmentRequest": {
            "type": "object",
            "required": [
                "approval_comment"
            ],
            "properties": {
                "approval_comment": {
                    "type": "string",
                    "maxLength": 512
                }
            }
        },
...
        "response.Body": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}